api/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
langgraph
langchain
python-dotenv
boto3>=1.34.0
httpx>=0.28.1,<0.29.0
python-multipart>=0.0.9
pytest>=8.0.0
pypdf>=4.2.0
langchain-openai
google-genai
langchain-google-genai
jinja2
aiofiles
requests-html
lxml_html_clean
tavily-python
daytona
